pyproject.toml
[project]
name = "lazy-pr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9",
    "pygithub>=2.9.1",
    "typer>=0.24.1",
]
